Update patch set 1
Patch Set 1:

Miguel,

the dzo way is an ugly hack. Changing the underlying conversion data is the _only_ clean solution.

Maybe the .cnv file needs to be integrated into some other file prior to building, I'm not entirely sure about that because I'm not exactly familiar with icu4c.

Patch-set: 1